Implemented biasing potential classes
