+++ openxps/__init__.py
@@ -3,6 +3,7 @@ Extended Phase-Space Methods with OpenMM
 """
 
 from ._version import __version__  # noqa: F401
+from .biasing import AdaptiveBiasingPotential, SplineBiasingPotential  # noqa: F401
 from .context import ExtendedSpaceContext  # noqa: F401
 from .extension_writer import ExtensionWriter  # noqa: F401
 from .extra_dof import ExtraDOF  # noqa: F401
@@ -11,4 +12,6 @@ __all__ = [
     "ExtendedSpaceContext",
     "ExtensionWriter",
     "ExtraDOF",
+    "SplineBiasingPotential",
+    "AdaptiveBiasingPotential",
 ]
